Update Socket Again V2

+++ Cloud/Domain/Model/SensorData.cs
@@ -10,7 +10,7 @@ public class SensorData
     public string Id { get; set; }
     public DateTime Timestamp { get; set; } = DateTime.UtcNow;
     public string MachineID { get; set; }
-    public int WaterTankLevel { get; set; }
+    public double WaterTankLevel { get; set; }
     public int MeasuredSoilMoisture { get; set; }
     public int AmountOfWatering { get; set; }
     public string PotId { get; set; }
